Try to fix test issue in github action by moving ping service code
